vgg16/requirements.txt
requests
oauth2client
flask
google-cloud-storage
